로그 추적기 적용
