+++ src/main/java/payment/example/ExampleApplication.java
@@ -3,10 +3,12 @@ package payment.example;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+
 @SpringBootApplication
 public class ExampleApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(ExampleApplication.class, args);
 	}
+
 }
