+++ src/main/java/payment/example/app/controller/dto/PaymentRequest.java
@@ -1,0 +1,18 @@
+package payment.example.app.controller.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class PaymentRequest {
+    private int amount;
+    private String impUid;
+    private String itemName;
+    private Long memberId;
+}
