+++ src/main/java/payment/example/app/service/dto/PaymentCancelRequest.java
@@ -1,0 +1,11 @@
+package payment.example.app.service.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+public class PaymentCancelRequest {
+    private String reason;
+    private String impUid;
+    private String amount;
+}
