+++ src/main/java/payment/example/common/domain/Address.java
@@ -1,0 +1,22 @@
+package payment.example.common.domain;
+
+import jakarta.persistence.Embeddable;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Embeddable
+@NoArgsConstructor
+public class Address {
+    private String city;
+    private String street;
+    private String zipcode;
+
+    @Builder
+    public Address(String city, String street, String zipcode) {
+        this.city = city;
+        this.street = street;
+        this.zipcode = zipcode;
+    }
+}
