+++ src/main/java/payment/example/common/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package payment.example.common.domain;
+
+public enum OrderStatus {
+    주문_성공, 주문_대기
+}
