+++ src/main/java/payment/example/common/exception/ItemStatusException.java
@@ -1,0 +1,18 @@
+package payment.example.common.exception;
+
+public class ItemStatusException extends RuntimeException {
+    public ItemStatusException() {
+    }
+
+    public ItemStatusException(String message) {
+        super(message);
+    }
+
+    public ItemStatusException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ItemStatusException(Throwable cause) {
+        super(cause);
+    }
+}
