+++ src/main/java/payment/example/common/exception/OutOfStockException.java
@@ -1,0 +1,18 @@
+package payment.example.common.exception;
+
+public class OutOfStockException extends RuntimeException {
+    public OutOfStockException() {
+    }
+
+    public OutOfStockException(String message) {
+        super(message);
+    }
+
+    public OutOfStockException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public OutOfStockException(Throwable cause) {
+        super(cause);
+    }
+}
