+++ src/main/java/payment/example/common/support/logtrace/LogTrace.java
@@ -1,0 +1,7 @@
+package payment.example.common.support.logtrace;
+
+public interface LogTrace {
+    TraceStatus start(String message, String ipAddr);
+
+    void end(TraceStatus traceStatus);
+}
