+++ src/main/java/payment/example/common/support/logtrace/TraceStatus.java
@@ -1,0 +1,19 @@
+package payment.example.common.support.logtrace;
+
+import lombok.Getter;
+
+@Getter
+public class TraceStatus {
+
+    private LogId logId;
+
+    private Long startTime;
+
+    private String message;
+
+    public TraceStatus(LogId logId, Long startTime, String message) {
+        this.logId = logId;
+        this.startTime = startTime;
+        this.message = message;
+    }
+}
