--- src/main/java/payment/example/controller/dto/PaymentRequest.java
@@ -1,16 +1,0 @@
-package payment.example.controller.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class PaymentRequest {
-    private int amount;
-    private String impUid;
-    private String itemName;
-    private Long memberId;
-}
