--- src/main/java/payment/example/domain/Address.java
@@ -1,22 +1,0 @@
-package payment.example.domain;
-
-import jakarta.persistence.Embeddable;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Embeddable
-@NoArgsConstructor
-public class Address {
-    private String city;
-    private String street;
-    private String zipcode;
-
-    @Builder
-    public Address(String city, String street, String zipcode) {
-        this.city = city;
-        this.street = street;
-        this.zipcode = zipcode;
-    }
-}
