--- src/main/java/payment/example/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package payment.example.domain;
-
-public enum OrderStatus {
-    주문_성공, 주문_대기
-}
