--- src/main/java/payment/example/exception/OutOfStockException.java
@@ -1,18 +1,0 @@
-package payment.example.exception;
-
-public class OutOfStockException extends RuntimeException {
-    public OutOfStockException() {
-    }
-
-    public OutOfStockException(String message) {
-        super(message);
-    }
-
-    public OutOfStockException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public OutOfStockException(Throwable cause) {
-        super(cause);
-    }
-}
