--- src/main/java/payment/example/service/dto/PaymentCancelRequest.java
@@ -1,11 +1,0 @@
-package payment.example.service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-public class PaymentCancelRequest {
-    private String reason;
-    private String impUid;
-    private String amount;
-}
