--- src/main/java/payment/example/support/Logger/HttpTransaction.java
@@ -1,6 +1,0 @@
-package payment.example.support.Logger;
-
-public enum HttpTransaction {
-    REQUEST,
-    RESPONSE
-}
