--- src/main/java/payment/example/support/Logger/LogTrace.java
@@ -1,9 +1,0 @@
-package payment.example.support.Logger;
-
-public interface LogTrace {
-    long start(LogStatus status);
-
-    void end(LogStatus status);
-
-    void exception();
-}
