+++ src/main/resources/application.yml
@@ -16,5 +16,4 @@ spring:
       hibernate:
         format_sql: true
 
-logging.level:
-  org.hibernate.SQL: debug
+
