+++ src/test/java/payment/example/domain/ItemTest.java
@@ -3,6 +3,7 @@ package payment.example.domain;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import payment.example.common.domain.Item;
 
 import static org.assertj.core.api.Assertions.*;
