+++ src/test/java/payment/example/domain/MemberTest.java
@@ -1,12 +1,12 @@
 package payment.example.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import payment.example.common.domain.Address;
+import payment.example.common.domain.Member;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 class MemberTest {
     @Test
