+++ src/test/java/payment/example/repository/ItemRepositoryTest.java
@@ -1,14 +1,12 @@
 package payment.example.repository;
 
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import payment.example.domain.Item;
-
-import static org.junit.jupiter.api.Assertions.*;
+import payment.example.common.domain.Item;
+import payment.example.app.repository.ItemRepository;
 
 @SpringBootTest
 @Transactional
